Changed administrator logics, added a few safe mechanisms for remote database usage

+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/CalculationVariableParametersController.java
@@ -3,6 +3,7 @@ package com.diploma.Diploma.Controllers.DatabaseInteractionForm;
 import com.diploma.Diploma.DataBase.Model.CalculationVariableParameters;
 import com.diploma.Diploma.DataBase.Repo.CalculationVariableParametersRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class CalculationVariableParametersController {
     }
 
     @PostMapping("/create")
-    public CalculationVariableParameters newCalculationVariableParameter(@RequestBody CalculationVariableParameters cvp){
-        return repo.save(cvp);
+    public ResponseEntity<CalculationVariableParameters> newCalculationVariableParameter(@RequestBody CalculationVariableParameters cvp){
+        Optional<CalculationVariableParameters> tmp = repo.findById(cvp.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(cvp));
+        }
     }
 
     @PostMapping("/update")
@@ -42,12 +48,18 @@ public class CalculationVariableParametersController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(cvp));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteCalculationVariableParameters(@RequestBody CalculationVariableParameters cvp){
-        repo.deleteById(cvp.getId());
+    public ResponseEntity<CalculationVariableParameters> deleteCalculationVariableParameters(@RequestBody CalculationVariableParameters cvp){
+        Optional<CalculationVariableParameters> tmp = repo.findById(cvp.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(cvp.getId());
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 }
