+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/EmergencyController.java
@@ -4,6 +4,7 @@ import com.diploma.Diploma.DataBase.Model.CloudCombustionMode;
 import com.diploma.Diploma.DataBase.Model.Emergency;
 import com.diploma.Diploma.DataBase.Repo.EmergencyRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class EmergencyController {
     }
 
     @PostMapping("/create")
-    public Emergency newEmergency(@RequestBody Emergency emergency){
-        return repo.save(emergency);
+    public ResponseEntity<Emergency> newEmergency(@RequestBody Emergency emergency){
+        Optional<Emergency> tmp = repo.findById(emergency.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(emergency));
+        }
     }
 
     @PostMapping("/update")
@@ -42,13 +48,20 @@ public class EmergencyController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(emergency));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteEmergency(@RequestBody Emergency emergency){
-        repo.deleteById(emergency.getId());
+    public ResponseEntity<Emergency> deleteEmergency(@RequestBody Emergency emergency){
+        Optional<Emergency> tmp = repo.findById(emergency.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(emergency.getId());
+            return ResponseEntity.ok().build();
+        } else{
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
+
     }
 }
