+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/EmergencyScenarioController.java
@@ -3,6 +3,7 @@ package com.diploma.Diploma.Controllers.DatabaseInteractionForm;
 import com.diploma.Diploma.DataBase.Model.EmergencyScenario;
 import com.diploma.Diploma.DataBase.Repo.EmergencyScenarioRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class EmergencyScenarioController {
     }
 
     @PostMapping("/create")
-    public EmergencyScenario newEmergencyScenario(@RequestBody EmergencyScenario emergencyScenario){
-        return repo.save(emergencyScenario);
+    public ResponseEntity<EmergencyScenario> newEmergencyScenario(@RequestBody EmergencyScenario es){
+        Optional<EmergencyScenario> tmp = repo.findById(es.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(es));
+        }
     }
 
     @PostMapping("/update")
@@ -42,12 +48,18 @@ public class EmergencyScenarioController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(emergency));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteEmergencyScenario(@RequestBody EmergencyScenario es){
-        repo.deleteById(es.getId());
+    public ResponseEntity<EmergencyScenario> deleteEmergencyScenario(@RequestBody EmergencyScenario es){
+        Optional<EmergencyScenario> tmp = repo.findById(es.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(es.getId());
+            return ResponseEntity.ok().build();
+        } else{
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 }
