+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/EnterpriseController.java
@@ -4,6 +4,7 @@ import com.diploma.Diploma.DataBase.Model.CloudCombustionMode;
 import com.diploma.Diploma.DataBase.Model.Enterprise;
 import com.diploma.Diploma.DataBase.Repo.EnterpriseRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class EnterpriseController {
     }
 
     @PostMapping("/create")
-    public Enterprise newEnterprise(@RequestBody Enterprise enterprise){
-        return repo.save(enterprise);
+    public ResponseEntity<Enterprise> newEnterprise(@RequestBody Enterprise enterprise){
+        Optional<Enterprise> tmp = repo.findById(enterprise.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(enterprise));
+        }
     }
 
     @PostMapping("/update")
@@ -42,12 +48,18 @@ public class EnterpriseController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(enterprise));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteEnterprise(@RequestBody Enterprise enterprise){
-        repo.deleteById(enterprise.getId());
+    public ResponseEntity<Enterprise> deleteEnterprise(@RequestBody Enterprise enterprise){
+        Optional<Enterprise> tmp = repo.findById(enterprise.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(enterprise.getId());
+            return ResponseEntity.ok().build();
+        } else{
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 }
