+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/EquipmentClassController.java
@@ -4,6 +4,7 @@ import com.diploma.Diploma.DataBase.Model.CloudCombustionMode;
 import com.diploma.Diploma.DataBase.Model.EquipmentClass;
 import com.diploma.Diploma.DataBase.Repo.EquipmentClassRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class EquipmentClassController {
     }
 
     @PostMapping("/create")
-    public EquipmentClass newEquipmentClass(@RequestBody EquipmentClass equipmentClass){
-        return repo.save(equipmentClass);
+    public ResponseEntity<EquipmentClass> newEquipmentClass(@RequestBody EquipmentClass equipmentClass){
+        Optional<EquipmentClass> tmp = repo.findById(equipmentClass.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(equipmentClass));
+        }
     }
 
     @PostMapping("/update")
@@ -42,12 +48,18 @@ public class EquipmentClassController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(equipmentClass));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteEquipmentClass(@RequestBody EquipmentClass equipmentClass){
-        repo.deleteById(equipmentClass.getId());
+    public ResponseEntity<EquipmentClass> deleteEquipmentClass(@RequestBody EquipmentClass equipmentClass){
+        Optional<EquipmentClass> tmp = repo.findById(equipmentClass.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(equipmentClass.getId());
+            return ResponseEntity.ok().build();
+        } else{
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 }
