+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/EquipmentInDepartmentController.java
@@ -4,6 +4,7 @@ import com.diploma.Diploma.DataBase.Model.CloudCombustionMode;
 import com.diploma.Diploma.DataBase.Model.EquipmentInDepartment;
 import com.diploma.Diploma.DataBase.Repo.EquipmentInDepartmentRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class EquipmentInDepartmentController {
     }
 
     @PostMapping("/create")
-    public EquipmentInDepartment newEquipmentInDepartment(@RequestBody EquipmentInDepartment equipmentInDepartment){
-        return repo.save(equipmentInDepartment);
+    public ResponseEntity<EquipmentInDepartment> newEquipmentInDepartment(@RequestBody EquipmentInDepartment equipmentInDepartment){
+        Optional<EquipmentInDepartment> tmp = repo.findById(equipmentInDepartment.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(equipmentInDepartment));
+        }
     }
 
     @PostMapping("/update")
@@ -42,12 +48,18 @@ public class EquipmentInDepartmentController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(equipmentInDepartment));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteEquipmentInDepartment(@RequestBody EquipmentInDepartment equipmentInDepartment){
-        repo.deleteById(equipmentInDepartment.getId());
+    public ResponseEntity<EquipmentInDepartment> deleteEquipmentInDepartment(@RequestBody EquipmentInDepartment equipmentInDepartment){
+        Optional<EquipmentInDepartment> tmp = repo.findById(equipmentInDepartment.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(equipmentInDepartment.getId());
+            return ResponseEntity.ok().build();
+        } else{
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 }
