+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/EventController.java
@@ -4,6 +4,7 @@ import com.diploma.Diploma.DataBase.Model.CloudCombustionMode;
 import com.diploma.Diploma.DataBase.Model.Event;
 import com.diploma.Diploma.DataBase.Repo.EventRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class EventController {
     }
 
     @PostMapping("/create")
-    public Event newEvent(@RequestBody Event event){
-        return repo.save(event);
+    public ResponseEntity<Event> newEvent(@RequestBody Event event){
+        Optional<Event> tmp = repo.findById(event.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(event));
+        }
     }
 
     @PostMapping("/update")
@@ -42,12 +48,18 @@ public class EventController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(event));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteEvent(@RequestBody Event event){
-        repo.deleteById(event.getId());
+    public ResponseEntity<Event> deleteEvent(@RequestBody Event event){
+        Optional<Event> tmp = repo.findById(event.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(event.getId());
+            return ResponseEntity.ok().build();
+        } else{
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 }
