+++ src/main/java/com/diploma/Diploma/Controllers/DatabaseInteractionForm/ExposureTypeController.java
@@ -4,6 +4,7 @@ import com.diploma.Diploma.DataBase.Model.CloudCombustionMode;
 import com.diploma.Diploma.DataBase.Model.ExposureType;
 import com.diploma.Diploma.DataBase.Repo.ExposureTypeRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,8 +33,13 @@ public class ExposureTypeController {
     }
 
     @PostMapping("/create")
-    public ExposureType newExposureType(@RequestBody ExposureType exposureType){
-        return repo.save(exposureType);
+    public ResponseEntity<ExposureType> newExposureType(@RequestBody ExposureType exposureType){
+        Optional<ExposureType> tmp = repo.findById(exposureType.getId());
+        if(tmp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else{
+            return ResponseEntity.ok(repo.save(exposureType));
+        }
     }
 
     @PostMapping("/update")
@@ -42,12 +48,18 @@ public class ExposureTypeController {
         if(tmp.isPresent()){
             return ResponseEntity.ok(repo.save(exposureType));
         } else{
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
     }
 
     @PostMapping("/delete")
-    public void deleteExposureType(@RequestBody ExposureType exposureType){
-        repo.deleteById(exposureType.getId());
+    public ResponseEntity<ExposureType> deleteExposureType(@RequestBody ExposureType exposureType){
+        Optional<ExposureType> tmp = repo.findById(exposureType.getId());
+        if(tmp.isPresent()){
+            repo.deleteById(exposureType.getId());
+            return ResponseEntity.ok().build();
+        } else{
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 }
